Digikam from stable : Fix wrong cast into kioslave. This bug crash digikam during album item thumbs rendering or stop it to the first item of each album
Nota : these patch have been already applied to trunk branch (0.9.0-svn)

CCMAIL: digikam-devel@kde.org
CCBUGS: 121905

svn path=/branches/stable/extragear/graphics/digikam/; revision=522589

+++ kioslave/digikamsearch.cpp
@@ -214,7 +214,7 @@ void kio_digikamsearch::special(const QByteArray& data)
             *os << dirid;
             *os << name;
             *os << date;
-            *os << stbuf.st_size;
+            *os << static_cast<size_t>(stbuf.st_size);
             *os << dims;
 
             count++;
